+++ datacube/index/memory/index.py
@@ -4,7 +4,9 @@
 # SPDX-License-Identifier: Apache-2.0
 import logging
 from threading import Lock
+from typing import Type
 
+from deprecat import deprecat
 from datacube.cfg import ODCEnvironment
 from datacube.index.memory._datasets import DatasetResource, LineageResource  # type: ignore
 from datacube.index.memory._fields import get_dataset_fields
@@ -13,6 +15,7 @@ from datacube.index.memory._products import ProductResource
 from datacube.index.memory._users import UserResource
 from datacube.index.abstract import AbstractIndex, AbstractIndexDriver, UnhandledTransaction
 from datacube.model import MetadataType
+from datacube.migration import ODC2DeprecationWarning
 from odc.geo import CRS
 
 _LOG = logging.getLogger(__name__)
@@ -26,6 +29,19 @@ class Index(AbstractIndex):
     """
     Lightweight in-memory index driver
     """
+    #   Metadata type support flags
+    supports_legacy = True
+    supports_eo3 = True
+    supports_nongeo = True
+
+    #   Database/storage feature support flags
+    supports_write = True
+
+    #   User managment support flags
+    supports_users = True
+
+    #   Lineage support flags
+    supports_lineage = True
 
     def __init__(self, env: ODCEnvironment) -> None:
         self._env = env
@@ -97,11 +113,16 @@ class Index(AbstractIndex):
 
 
 class MemoryIndexDriver(AbstractIndexDriver):
-    @staticmethod
-    def connect_to_index(config_env: ODCEnvironment, application_name=None, validate_connection=True):
-        return Index.from_config(config_env, application_name, validate_connection)
+    @classmethod
+    def index_class(cls) -> Type[AbstractIndex]:
+        return Index
 
     @staticmethod
+    @deprecat(
+        reason="The 'metadata_type_from_doc' static method has been deprecated. "
+               "Please use the 'index.metadata_type.from_doc()' instead.",
+        version='1.9.0',
+        category=ODC2DeprecationWarning)
     def metadata_type_from_doc(definition: dict) -> MetadataType:
         """
         :param definition: